+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 v 1.0
+  - Do not quit and issue a warning if changes not saved. #3
   - Persist bookmark list to disk. #1